+++ CHANGES.rst
@@ -11,6 +11,13 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 
 - Update dependencies to the latest releases for each supported Python version.
 
+- Fix cookie path parameter handling:
+  If the cookie path value contains only characters allowed
+  in URL path components (with the exception of ``;``) including ``%``,
+  it is used as is; otherwise, it is quoted using Python's
+  ``urllib.parse.quote``
+  (`#1052 <https://github.com/zopefoundation/Zope/issues/1052>`_).
+
 
 4.8.2 (2022-06-01)
 ------------------