+++ CHANGES.rst
@@ -12,8 +12,10 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 - Update dependencies to the latest releases for each supported Python version.
 
 - Fix cookie path parameter handling:
-  If the cookie path value contains only characters allowed
-  in URL path components (with the exception of ``;``) including ``%``,
+  If the cookie path value contains ``%`` it is assumed to be
+  fully quoted and used as is;
+  if it contains only characters allowed (unquoted)
+  in an URL path (with the exception of ``;``),
   it is used as is; otherwise, it is quoted using Python's
   ``urllib.parse.quote``
   (`#1052 <https://github.com/zopefoundation/Zope/issues/1052>`_).