+++ CHANGES.rst
@@ -14,7 +14,7 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 
 - Update to newest compatible versions of dependencies.
 
-- Add preliminary support for Python 3.11 (as of 3.11.0b3).
+- Add preliminary support for Python 3.11 (as of 3.11.0b5).
 
 - Fix cookie path parameter handling:
   If the cookie path value contains ``%`` it is assumed to be